feat(formatter): XML formatter
Copy from `collective.xmltestrunner`,
with `six` code paths removed.
